make method in SyntheticRecordType abstract and move implementations to its children
