+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/metadata/JoinedRecordType.java
@@ -21,14 +21,24 @@
 package com.apple.foundationdb.record.metadata;
 
 import com.apple.foundationdb.annotation.API;
+import com.apple.foundationdb.async.AsyncUtil;
 import com.apple.foundationdb.record.RecordMetaData;
 import com.apple.foundationdb.record.RecordMetaDataProto;
 import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
 import com.apple.foundationdb.record.metadata.expressions.LiteralKeyExpression;
+import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStore;
+import com.apple.foundationdb.record.provider.foundationdb.FDBStoredRecord;
+import com.apple.foundationdb.record.provider.foundationdb.FDBSyntheticRecord;
+import com.apple.foundationdb.record.provider.foundationdb.RecordDoesNotExistException;
+import com.apple.foundationdb.tuple.Tuple;
 import com.google.protobuf.Descriptors;
+import com.google.protobuf.Message;
 
 import javax.annotation.Nonnull;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * A <i>synthetic</i> record type representing the indexable result of <em>joining</em> stored records.
@@ -110,6 +120,32 @@ public class JoinedRecordType extends SyntheticRecordType<JoinedRecordType.JoinC
     public List<Join> getJoins() {
         return joins;
     }
+
+    @Nonnull
+    @Override
+    @API(API.Status.INTERNAL)
+    public CompletableFuture<FDBSyntheticRecord> loadByPrimaryKeyAsync(FDBRecordStore store, Tuple primaryKey) {
+        int nconstituents = getConstituents().size();
+        final Map<String, FDBStoredRecord<? extends Message>> constituentValues = new ConcurrentHashMap<>(nconstituents);
+        final CompletableFuture<?>[] futures = new CompletableFuture<?>[nconstituents];
+        for (int i = 0; i < nconstituents; i++) {
+            final SyntheticRecordType.Constituent constituent = getConstituents().get(i);
+            final Tuple constituentKey = primaryKey.getNestedTuple(i + 1);
+            if (constituentKey == null) {
+                futures[i] = AsyncUtil.DONE;
+            } else {
+                futures[i] = store.loadRecordAsync(constituentKey).thenApply(rec -> {
+                    if (rec == null) {
+                        throw new RecordDoesNotExistException("constituent record not found: " + constituent.getName());
+                    }
+                    constituentValues.put(constituent.getName(), rec);
+                    return null;
+                });
+            }
+        }
+        return CompletableFuture.allOf(futures).thenApply(vignore -> FDBSyntheticRecord.of(this, constituentValues));
+    }
+
 
     @Nonnull
     public RecordMetaDataProto.JoinedRecordType toProto() {
